Fixed zero-roots case in bezier_above

+++ src/utils.h
@@ -13,7 +13,7 @@ void swap(float* a, float* b);
 float gammln(const float xx);
 float factrl(const int n);
 float factln(const int n);
-float bico(const int n, const int k);
+int bico(const int n, const int k);
 int up_power(const int x, const int up);
   
 const float PI;
